requirements.txt
numpy==1.22.3
Pillow==9.1.1
Pillow==10.1.0
scipy==1.8.1
torch==2.1.1
torchvision==0.16.1
